+++ openpyxl/workbook/workbook.py
@@ -88,6 +88,7 @@ class Workbook(object):
         self.rels = RelationshipList()
         self.calculation = CalcProperties()
         self.views = [BookView()]
+        self._volatile_deps = None
 
 
     def _setup_styles(self):