+++ doc/changes.rst
@@ -7,7 +7,8 @@ New Features
 
 * Support Active X elements
 * Support Form Controls
-
+* Support for documents with volatile dependencies
+ 
 
 Deprecations
 ------------